+++ chaosaws/ec2_os/constants.py
@@ -1,0 +1,9 @@
+# OS Types
+OS_WINDOWS = "Windows"
+OS_LINUX = "Linux"
+
+# script name
+BURN_CPU = "cpu_stress_test"
+BURN_IO = "burn_io"
+FILL_DISK = "fill_disk"
+NETWORK_UTIL = "network_advanced"